+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.1.1 (Next)
 
 * [#2](https://github.com/dblock/slack-ruby-client/pull/2): `Slack::RealTime::Socket` now pings frames every 30s, as recommended by Slack - [@samdoiron](https://github.com/samdoiron).
+* Exposed `Slack::RealTime::Client.url`, `team`, `self`, `users`, `channels`, `groups`, `ims` and `bots` - [@dblock](https://github.com/dblock).
+
 * Your contribution here.
 
 ### 0.1.0 (7/25/2015)